feat: completar ejercicio autonomo 5 en 06_archivos_y_modulos

+++ 06_archivos_y_modulos/autonomos/ejercicio5/utiles.py
@@ -1,10 +1,17 @@
 # utiles.py
 # Módulo donde se almacena funciones útiles para leer y escribir archivos
 
+import os
+
 # TODO 1: Función que lee el contenido del archivo y las líneas de texto que contiene
 def leer_lineas(ruta):
+
+    if os.path.exists(ruta):
+        print(f"\nSobrescribiendo archivo {ruta}...")
+
+    print("\nLeyendo archivo... ")
     try:
-        print("Leyendo archivo... \n")
+            
         with open(ruta, 'r', encoding="utf-8") as archivo:
 
             contenido = archivo.read()
@@ -18,22 +25,38 @@ def leer_lineas(ruta):
 
         print(f"Archivo {ruta} corregido con éxito ✅")
 
+    # Agregar mensaje de error usando FileNotFoundError si no encuentra el archivo
     except FileNotFoundError:
         print("❌ Archivo no encontrado.\n")
+    # Agregar mensaje de error usando FileExistsError si existe el archivo
+    except FileExistsError:
+        print(f"❌ El archivo {ruta} Ya existe ")
+    # Agregar mensaje de error usando Exception si ha surgido otro tipo de error
     except Exception as e:
         print(f"\n❌ Error al leer el archivo: {e}")
 
 # TODO 2: Función para guardar las líneas a un archivo específico
 def guardar_lineas(ruta, lineas):
-    try:
-        print("Escribiendo archivo...\n")
-        with open(ruta, 'w', encoding="utf-8") as archivo:
 
-            for linea in lineas:
-                archivo.write(f"\n{linea}")
+    print("\n Escribiendo archivo...")
+    if os.path.exists(ruta):
+        print(f"❌ El archivo {ruta} Ya existe ")
+    
+    else:
+        try:
+            
+            with open(ruta, 'w', encoding="utf-8") as archivo:
+
+                for linea in lineas:
+                    archivo.write(f"\n{linea}")
         
-        print(f"Archivo {ruta} creado con éxito ✅")
-    except FileNotFoundError:
-        print("❌ Archivo no encontrado.\n")
-    except Exception as e:
-        print(f"\n❌ Error al escribir el archivo: {e}")
+            print(f"Archivo {ruta} creado con éxito ✅")
+        # Agregar mensaje de error usando FileNotFoundError si no encuentra el archivo
+        except FileNotFoundError:
+            print("❌ Archivo no encontrado.\n")
+        # Agregar mensaje de error usando FileExistsError si existe el archivo
+        except FileExistsError:
+            print(f"❌ El archivo {ruta} Ya existe ")
+        # Agregar mensaje de error usando Exception si ha surgido otro tipo de error
+        except Exception as e:
+            print(f"\n❌ Error al escribir el archivo: {e}")
